Record cassete for possible NetSuite bug
Some options ids returned by webservice and the names we see on the
NetSuite browser UI don't match. e.g.

This is a matrix list of options:

    <platformCore:customField internalId="4627" scriptId="custitem18" xsi:type="platformCore:MultiSelectCustomFieldRef">
      <platformCore:value internalId="10" typeId="529">
        <platformCore:name>Granite</platformCore:name>
      </platformCore:value>
      <platformCore:value internalId="8" typeId="529">
        <platformCore:name>Moss</platformCore:name>
      </platformCore:value>
      <platformCore:value internalId="9" typeId="529">
        <platformCore:name>Retro Khaki</platformCore:name>
      </platformCore:value>
    </platformCore:customField>

This is the matrix option list in one of that child matrix above:

    <listAcct:matrixOptionList>
      <listAcct:matrixOption internalId=\"4627\" scriptId=\"custitem18\">
        <platformCore:value internalId=\"8\" typeId=\"529\"/>
      </listAcct:matrixOption>
      <listAcct:matrixOption internalId=\"4629\" scriptId=\"custitem19\">
      <platformCore:value internalId=\"1\" typeId=\"530\"/>
      </listAcct:matrixOption>
    </listAcct:matrixOptionList>

According to the parent item the option type_id "529" and internal id
"8" is however we looking at that specific item on NetSuite UI we see
actually "Granite"

looks like a netsuite issue or I'm just too tired